Moving explicit interfaces to declared typing files
