+++ webapp/.vscode/settings.json
@@ -1,0 +1,7 @@
+{
+    "prettier.eslintIntegration": true,
+    "prettier.printWidth": 120,
+    "prettier.semi": false,
+    "prettier.singleQuote": false,
+    "prettier.tabWidth": 4,
+}
