+++ webapp/src/Models/types/Error.d.ts
@@ -1,4 +1,4 @@
-export interface AppError {
+interface AppError {
     code?: number
     message: string
 }
