+++ webapp/src/Models/types/GlobalStatsData.d.ts
